// portal-api/src/model/kanel/public/UserServiceCapability.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { GenericServiceCapabilityId } from './GenericServiceCapability';
import type { SubscriptionCapabilityId } from './SubscriptionCapability';
import type { UserServiceId } from './UserService';

/** Identifier type for public.UserService_Capability */
export type UserServiceCapabilityId = string & {
  __brand: 'UserServiceCapabilityId';
};

/** Represents the table public.UserService_Capability */
export default interface UserServiceCapability {
  id: UserServiceCapabilityId;

  user_service_id: UserServiceId | null;

  generic_service_capability_id: GenericServiceCapabilityId | null;

  subscription_capability_id: SubscriptionCapabilityId | null;
}

/** Represents the initializer for the table public.UserService_Capability */
export interface UserServiceCapabilityInitializer {
  /** Default value: gen_random_uuid() */
  id?: UserServiceCapabilityId;

  user_service_id?: UserServiceId | null;

  generic_service_capability_id?: GenericServiceCapabilityId | null;

  subscription_capability_id?: SubscriptionCapabilityId | null;
}

/** Represents the mutator for the table public.UserService_Capability */
export interface UserServiceCapabilityMutator {
  id?: UserServiceCapabilityId;

  user_service_id?: UserServiceId | null;

  generic_service_capability_id?: GenericServiceCapabilityId | null;

  subscription_capability_id?: SubscriptionCapabilityId | null;
}